fix: resolve all golangci-lint errors for CI
- errcheck: add `_ =` prefix to unchecked pgContainer.Terminate and tdb.Container.Terminate calls in testdb.go
- errcheck: add `_, _ =` prefix to unchecked fmt.Sscanf calls in systemmetadata/service.go
- unconvert: remove unnecessary int32() conversions in assets/service.go and string() conversions in notifications/service.go
- unused: remove unused parseTimeBucket function and its associated imports from timeline/server.go
- copyloopvar: remove redundant `tc := tc` loop variable copies in assets/metadata_test.go (Go 1.22+)
- gocritic ifElseChain: convert if-else chain to switch statement in faces/service.go
- gocritic assignOp: use `+=` instead of `= addr + ...` in server/server.go
- gofumpt: apply gofumpt formatting to maintenance/service.go, config_test.go, and timeline/server.go
- gosimple S1009: omit nil check before len() for nil map in notifications/server.go
- ineffassign: replace ctx with _ in tracer.Start calls where ctx is not used after in maintenance, plugin, queue, and workflow services
- nakedret: convert named return in parseRedisURL to explicit return values in server/server.go
- gosec G306: fix file permission from 0644 to 0600 in config_test.go

+++ internal/assets/service.go
@@ -1044,11 +1044,11 @@ func (s *Service) convertToAssetInfo(asset sqlc.Asset, thumbnails []AssetThumbna
 			info.Metadata.Size = exif.FileSizeInByte.Int64
 		}
 		if exif.ExifImageWidth.Valid {
-			width := int32(exif.ExifImageWidth.Int32)
+			width := exif.ExifImageWidth.Int32
 			info.Metadata.Width = &width
 		}
 		if exif.ExifImageHeight.Valid {
-			height := int32(exif.ExifImageHeight.Int32)
+			height := exif.ExifImageHeight.Int32
 			info.Metadata.Height = &height
 		}
 		if exif.Make.Valid {
@@ -1067,7 +1067,7 @@ func (s *Service) convertToAssetInfo(asset sqlc.Asset, thumbnails []AssetThumbna
 			info.Metadata.FocalLength = &exif.FocalLength.Float64
 		}
 		if exif.Iso.Valid {
-			iso := int32(exif.Iso.Int32)
+			iso := exif.Iso.Int32
 			info.Metadata.ISO = &iso
 		}
 		if exif.ExposureTime.Valid {
